+++ cap/cap-impl/src/main/java/org/restcomm/protocols/ss7/cap/MessageImpl.java
@@ -37,11 +37,11 @@ import io.netty.util.ReferenceCountUtil;
 public abstract class MessageImpl implements CAPMessage {
 	private static final long serialVersionUID = 1L;
 
-	private long invokeId;
+	private int invokeId;
     private CAPDialog capDialog;
     private ByteBuf originalBuffer;
     
-    public long getInvokeId() {
+    public int getInvokeId() {
         return this.invokeId;
     }
 
@@ -49,7 +49,7 @@ public abstract class MessageImpl implements CAPMessage {
         return this.capDialog;
     }
 
-    public void setInvokeId(long invokeId) {
+    public void setInvokeId(int invokeId) {
         this.invokeId = invokeId;
     }
