+++ cap/cap-impl/src/main/java/org/restcomm/protocols/ss7/cap/service/sms/CAPServiceSmsImpl.java
@@ -118,9 +118,9 @@ public class CAPServiceSmsImpl extends CAPServiceBaseImpl implements CAPServiceS
 
     @Override
     public void processComponent(ComponentType compType, OperationCode oc, CAPMessage parameter, CAPDialog capDialog,
-            Long invokeId, Long linkedId) throws CAPParsingComponentException {
+            Integer invokeId, Integer linkedId) throws CAPParsingComponentException {
 
-    	Long ocValue = oc.getLocalOperationCode();
+    	Integer ocValue = oc.getLocalOperationCode();
         if (ocValue == null)
             new CAPParsingComponentException("", CAPParsingComponentExceptionReason.UnrecognizedOperation);
